Don't run scripts as root when it's not needed to
Most scripts don't have to be run as root and so they shouldn't be run
as root. Display an error message and bail out if these are lauched by
the root user.

+++ 02-arrangepackages.sh
@@ -1,5 +1,10 @@
 #/bin/sh
 
+if [ "$UID" -eq "0" ]; then
+	echo "Don't run this script as root"
+	exit 1
+fi
+
 rm -rf iso
 rm -rf temp
 mkdir -p iso/salix/{kernels,core,basic,full,settings}
