+++ 04-md5andlist.sh
@@ -1,5 +1,10 @@
 #/bin/sh
 
+if [ "$UID" -eq "0" ]; then
+	echo "Don't run this script as root"
+	exit 1
+fi
+
 rm -f iso/CHECKSUMS.md5 iso/PACKAGELIST iso/PACKAGELIST-TEMP
 
 for pkg in `find ./iso/salix -type f -name '*.t[gx]z' -print`
